# chore: improve `/healthz` endpoint performance